#23 fixing locking mechanism

+++ src/main/java/de/cismet/cismap/cidslayer/CidsBeanLocker.java
@@ -73,8 +73,13 @@ public class CidsBeanLocker {
             final String userString = NbBundle.getMessage(
                     CidsBeanLocker.class,
                     "CidsLayerLocker.lock(CidsBean).userString",
-                    SessionManager.getSession().getUser().toString());
-
+                    SessionManager.getSession().getUser().getName());
+            
+            if (bean.getMetaObject().getMetaClass().getID() == -1) {
+                //object with id -1 are new objects, which cannot be locked, because they do not exists on the server
+                return new FakeLock();
+            }
+            
             // Check, if the lock already exists
             final String query = String.format(
                     LOCK_QUERY,
@@ -138,7 +143,7 @@ public class CidsBeanLocker {
             final String userString = NbBundle.getMessage(
                     CidsBeanLocker.class,
                     "CidsLayerLocker.lock(CidsBean).userString",
-                    SessionManager.getSession().getUser().toString());
+                    SessionManager.getSession().getUser().getName());
 
             // Check, if the lock already exists
             final String query = String.format(
@@ -194,6 +199,9 @@ public class CidsBeanLocker {
      */
     public void unlock(final CidsBean bean) throws Exception {
         try {
+            if(bean instanceof FakeLock) {
+                return;
+            }
             bean.delete();
             bean.persist();
         } catch (Exception e) {
@@ -225,5 +233,8 @@ public class CidsBeanLocker {
         }
 
         return lockMc;
+    }
+    
+    private class FakeLock extends CidsBean {
     }
 }
